+++ beasts_and_bards/android/app/src/main/kotlin/com/beastsandbards/android/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.beasts_and_bards
+package com.beastsandbards.android
 
 import io.flutter.embedding.android.FlutterActivity
