+++ beasts_and_bards/web/index.html
@@ -29,7 +29,7 @@
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>beasts_and_bards</title>
+  <title>beastsandbards</title>
   <link rel="manifest" href="manifest.json">
 
   <script>
