+++ beasts_and_bards/windows/runner/Runner.rc
@@ -90,12 +90,12 @@ BEGIN
         BLOCK "040904e4"
         BEGIN
             VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "beasts_and_bards" "\0"
+            VALUE "FileDescription", "beastsandbards" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "beasts_and_bards" "\0"
+            VALUE "InternalName", "beastsandbards" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2022 com.example. All rights reserved." "\0"
             VALUE "OriginalFilename", "beasts_and_bards.exe" "\0"
-            VALUE "ProductName", "beasts_and_bards" "\0"
+            VALUE "ProductName", "beastsandbards" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
